I'm fed up with trying to get it to work like I want.
Preferably BotDebug window would be a toolbar that is a child of the window with the GLCanvas, and be dockable to either the top or bottom of the screen.

Also, I'd like teh window with the GLCanvas to have restore and minimize buttons when it's maximized.  In the future I'd like to have some windows with graphs, etc. that are mdichildren in the same way that that window is.