Worked some more on the GUIBotDebug control.  Very close to working now.  Right clicking opens window.
I'm having difficult having the control resize itself correctly after the user hides one of the frames.

Before it can be finished, the engine needs to be put in a seperate thread from the rest of the GUI, and I'm having trouble figuring out FOX's apparent built in threading support.  I'm loathe to use Windows specific code for it, even for debugging purposes.