Fixed a major bug in the senses module.
Ran across a-- well, something or other.  Maybe a bug.  In Sprotiels's new tie routines.  I've marked it.  It causes an error when I try to run the source.  I need to know if this is a compiler specific problem or not.

Speaking of which, MSVC 6.0 can be downloaded from the tools section of the FTP.