Relatively minor work done on robot physics.
Presently there is some elusive bug I have been unable to track down.  In my simulations, when veggies get 6000 energy to reproduce, it causes all of their kids to [either think they have or actually have] 6000 energy, which causes a cascade effect with millions of little baby plants being born every cycle.

It would appear that somehow information of the parent's nrg levels before reproduction is somehow being given to the children but is properly updated for the parent.  Or perhaps the reverse is true, but I'm rather certain it's the children that continue to reproduce.