Fixed what I thought was an issue with shot collisions, but was in reality a problem of addressing the robot array.
For future reference to myself and others, when using a for loop to cycle through all the robots, it must be in the form of:

for(int x = 0; x <= MaxRobs; x++)

or an equivelant.  I had it doing x < MaxRobs in many routines, which was missing the last robot.

My 'phantom' shot collision problem was actually shots colliding with a robot that wasn't being drawn.